Update | Remove the 1 from Integer - line 7

+++ part3/hbnb/app/models/review.py
@@ -4,7 +4,7 @@ from app.extensions import db
 class Review(BaseModel):
     __tablename__= 'reviews'
     text = db.Column(db.String(100), nullable=False)
-    rating = db.Column(db.Integer(1), nullable=False)
+    rating = db.Column(db.Integer, nullable=False)
 
     def update(self, review_data):
         """
